requirements.txt
discord.py==1.6.0
feedparser==6.0.2
mysql-connector-python==8.0.23
psycopg2==2.8.6
python-dotenv==0.15.0

